[ANDROSDK-1452] Refactor fileResources post call (old/new importer)
