--- core/src/main/java/org/hisp/dhis/android/core/fileresource/internal/FileResourceStoreImpl.java
@@ -1,74 +1,0 @@
-/*
- *  Copyright (c) 2004-2021, University of Oslo
- *  All rights reserved.
- *
- *  Redistribution and use in source and binary forms, with or without
- *  modification, are permitted provided that the following conditions are met:
- *  Redistributions of source code must retain the above copyright notice, this
- *  list of conditions and the following disclaimer.
- *
- *  Redistributions in binary form must reproduce the above copyright notice,
- *  this list of conditions and the following disclaimer in the documentation
- *  and/or other materials provided with the distribution.
- *  Neither the name of the HISP project nor the names of its contributors may
- *  be used to endorse or promote products derived from this software without
- *  specific prior written permission.
- *
- *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- *  ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- *  WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- *  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- *  ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- *  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- *  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- *  ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.hisp.dhis.android.core.fileresource.internal;
-
-import android.database.Cursor;
-
-import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
-import org.hisp.dhis.android.core.arch.db.querybuilders.internal.SQLStatementBuilderImpl;
-import org.hisp.dhis.android.core.arch.db.stores.binders.internal.StatementBinder;
-import org.hisp.dhis.android.core.arch.db.stores.internal.IdentifiableDataObjectStoreImpl;
-import org.hisp.dhis.android.core.fileresource.FileResource;
-import org.hisp.dhis.android.core.fileresource.FileResourceTableInfo;
-
-import kotlin.jvm.functions.Function1;
-
-public final class FileResourceStoreImpl extends IdentifiableDataObjectStoreImpl<FileResource> {
-
-    private static final StatementBinder<FileResource> BINDER = (o, w) -> {
-        w.bind(1, o.uid());
-        w.bind(2, o.name());
-        w.bind(3, o.created());
-        w.bind(4, o.lastUpdated());
-        w.bind(5, o.contentType());
-        w.bind(6, o.contentLength());
-        w.bind(7, o.path());
-        w.bind(8, o.syncState());
-    };
-
-    private FileResourceStoreImpl(DatabaseAdapter databaseAdapter,
-                                  SQLStatementBuilderImpl builder,
-                                  StatementBinder<FileResource> binder,
-                                  Function1<Cursor, FileResource> objectFactory) {
-        super(databaseAdapter, builder, binder, objectFactory);
-    }
-
-    public static FileResourceStoreImpl create(DatabaseAdapter databaseAdapter) {
-        SQLStatementBuilderImpl statementBuilder = new SQLStatementBuilderImpl(
-                FileResourceTableInfo.TABLE_INFO.name(),
-                FileResourceTableInfo.TABLE_INFO.columns());
-
-        return new FileResourceStoreImpl(
-                databaseAdapter,
-                statementBuilder,
-                BINDER,
-                FileResource::create
-        );
-    }
-}
